🖊️ Trim leading and trailing spaces in assignments (#5936)
With this PR we explicitly trim leading and trailing spaces in assignments in levels 2-11. Note that in these levels the assignment and list assignment statements do not quotes. The program `a is three spaces   ` used to be transpiled to `a = 'three spaces   '` but is now transpiled to `a = 'three spaces'`. Similarly, the list `a is  cat , dog ,  chicken ` is now transpiled to `a = ['cat', 'dog', 'chicken']`.

Fixes #5684

**How to test**
Ensure that the following snippet yields `[4]` twice in level 9. Ensure that there is no grammar ambiguity. 
```
i = 4   # met comment
print '[' i ']'
if 1 = 1
    j = 4    # met comment
    print '[' j ']'
```