💻 New public adventures page  (#6092)
This PR changes the UI of the teachers "public adventures" page. It does away with the tabs and replaces it with a scrollable list on the left side. It's also faster and doesn't rely on all requests coming in to a single Python process (which will not be true on Heroku).

Behavioral changes w.r.t the previous version:

* A cloned public adventure starts out as private (used to be public). With the old behavior, after cloning, we would just have 2 public copies of the exact same adventure, which seems to not add any value. If a teacher changes it (and puts `print my boss is a poopy-head!` in the adventure), those changes are inadverently public  It seems to make more sense that a user deciding to republish and adventure should be a conscious decision.
* It's no longer possible to see adventures from all languages at the same time: selecting a language is required. I figure that real users are interested in a single language only, and if you are a Spanish speaker adventures in Chinese or Telugu don't help you much.

This is what it currently looks like:

<img width="1538" alt="image" src="https://github.com/user-attachments/assets/0f0e9340-d1a9-41aa-8c28-2f24c718037e" />


**How to test**

Log in as a teacher, and search and use the interface. Use clone, use flag, use delete if you are `teacher1`.


Co-Authored-By: boryanagoncharenko <boryanagoncharenko@users.noreply.github.com>