💻 Show loading spinner when opening code page (#6347)
When users open our page, and they are on a slow connection, there's no indiciation for them to know that the page is loading. Therefore it seems that page is unresponsive and there's an error.

This PR addresses that by not showing content directly to the user when the code page loads, but rather a stub of an adventure with an animation, indicating that the javascript is loading. To prevent the behaviour of clicking a button and showing an error, I disabled the buttons while the page loads.

Here's a video of how it looks:

https://github.com/user-attachments/assets/aea486b0-ca10-4a68-b882-8fbfd2a4b912

Fixes #6312

**How to test**

Open your browser's dev tools by pressing F12. Then go to the network tab and click on disable caching, on Throttling click on slow 4G. Reload the page and after a bit you should see the new spinner.

![image](https://github.com/user-attachments/assets/5a907934-7bce-4853-943c-f280541a457e)



Co-Authored-By: boryanagoncharenko <boryanagoncharenko@users.noreply.github.com>
Co-Authored-By: Kevin C <Kache@users.noreply.github.com>